Refactoring the code, more validations, inlcuding some Data Transfer Objects.

+++ src/main/java/firstwebjavaproject/javawebproject/services/PlayerServiceImpl.java
@@ -1,6 +1,8 @@
 package firstwebjavaproject.javawebproject.services;
 
 import firstwebjavaproject.javawebproject.model.Player;
+import firstwebjavaproject.javawebproject.model.PlayerAward;
+import firstwebjavaproject.javawebproject.repository.PlayerAwardRepository;
 import firstwebjavaproject.javawebproject.repository.PlayerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -12,6 +14,8 @@ import java.util.Optional;
 public class PlayerServiceImpl implements PlayerService{
     @Autowired
     private PlayerRepository playerRepository;
+    @Autowired
+    private PlayerAwardRepository playerAwardRepository;
     @Override
     public void savePlayer(Player player) {
         playerRepository.save(player);
@@ -29,6 +33,29 @@ public class PlayerServiceImpl implements PlayerService{
 
     @Override
     public void deleteById(Long id) {
-        playerRepository.deleteById(id);
+        Optional<Player> optionalPlayer = playerRepository.findById(id);
+        if(optionalPlayer.isPresent()){
+            Player player = optionalPlayer.get();
+            for ( PlayerAward playerAward: player.getPlayerAwards()) {
+            playerAwardRepository.delete(playerAward);
+            }
+            playerRepository.deleteById(id);
+        }
+    }
+
+    @Override
+    public List<Player> getPlayersById(List<Integer> playerIds) {
+      return playerRepository.findByIdIn(playerIds);
+    }
+
+    @Override
+    public Player getPlayerById(Long id) {
+      Optional<Player> player = playerRepository.findById(id);
+      if (player.isPresent()){
+          return player.get();
+      }
+      else{
+          throw new RuntimeException("Player not found with id: " + id);
+      }
     }
 }
